fixed cannot auth jwt completed

+++ internal/users/users.go
@@ -9,7 +9,7 @@ import (
 )
 
 type User struct {
-	ID              int64  `json:"id"`
+	ID              float64  `json:"id"`
 	Username        string `json:"username"`
 	Email           string `json:"email"`
 	FirstName       string `json:"first_name"`
